feat: introduce EnrollmentEmailTemplatesRetriever
Render availables email templates interpolated
for a specific enrollment which can be customized by an instructor.

There's 4 kinds of email:

* review_application
* validate_application
* refuse_application
* notify

By default it renders the default view with the enrollment mailer
layout.

If for the target api there's a view available it takes this view,
without the enrollment mailer layout.