refactor: make bridges more rails way
This commit move bridges in a `app/bridges` folder and introduce an
application bridge object with the interface to implement.

The old hierarchy may lead to uninitialized constant errors. For
example, the file `app/services/bridge/aidants_connect_bridge.rb` should
be named `Bridge::AidantsConnectBridge` with this hierarchy and the
autoload process of Rails (which, in development and test raises
errors). It may lead on some production configuration to uninitialized
constant error too (which is not safe)

There is no production code change, it should be safe.