feat: can customize backends enrollment emails layout inclusion
This commit introduces a new configuration in config/data_providers.yml:
`skip_layout`, which is a boolean.

Thanks to this configuration, it is possible to not to use the default
layout for a specific event.

If `skip_layout` is true on an event, the template is sent without
layout.

This commit implements this behaviour for 2 templates:

* send_application for api_entreprise
* create_application for api_entreprise